requirements.txt
flask==3.0.0
openai==1.60.0
python-dotenv==1.0.0
pdf2image==1.16.3
PyPDF2==3.0.1
Pillow==10.1.0 
pydub==0.25.1
httpx>=0.23.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27